fix : 불필요한 state 정리 및 ui layer 레이아웃 분리
